Generacion de token y proteccion de los endpoints

+++ src/main/java/com/procesos/negocio/fabian/controllers/UsuarioController.java
@@ -4,6 +4,7 @@ import com.procesos.negocio.fabian.models.Usuario;
 import com.procesos.negocio.fabian.repository.UsuarioRepository;
 import com.procesos.negocio.fabian.services.UsuarioService;
 import com.procesos.negocio.fabian.services.UsuarioServiceImpl;
+import com.procesos.negocio.fabian.utils.JWTUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -20,47 +21,61 @@ public class UsuarioController {
 
     @Autowired
    private UsuarioService usuarioService;
+    @Autowired
+    private JWTUtil jwtUtil;
     @GetMapping(value = "/usuario/{id}")
-    public ResponseEntity getUsuario(@PathVariable Long id){
+    public ResponseEntity getUsuario(@PathVariable Long id, @RequestHeader(value="Authorization") String token){
+        if(jwtUtil.getKey(token) == null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token no valido");
+        }
         return usuarioService.getUserById(id);
-
-       /* Usuario usuario= new Usuario();
-        usuario.setId(id);
-        usuario.setNombre("fabian");
-        usuario.setApellidos("Rincón Chinchilla");
-        usuario.setDocumento("1064836389");
-        usuario.setFechaNacimiento(new Date(2004,7,14));
-        usuario.setDireccion("kdx-o10-310");
-        usuario.setTelefono("3144454761");*/
-
     }
     @PostMapping("/usuario")
     public ResponseEntity crearUsuario(@Valid @RequestBody  Usuario usuario){
         return usuarioService.createUser(usuario);
     }
     @GetMapping(value="/usuarios")
-    public ResponseEntity listarUsuarios(){
-        return usuarioService.allUsers();
+    public ResponseEntity listarUsuarios(@RequestHeader(value="Authorization") String token){
+        if(jwtUtil.getKey(token) != null){
+            return usuarioService.allUsers();
+        }
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token no valido");
+
     }
 
     @GetMapping(value="/usuarios/{nombre}/{apellidos}")
-    public ResponseEntity listarPorNombreApellidos(@PathVariable  String nombre, @PathVariable String apellidos ){
+    public ResponseEntity listarPorNombreApellidos(@PathVariable  String nombre, @PathVariable String apellidos, @RequestHeader(value="Authorization") String token ){
+        if(jwtUtil.getKey(token) == null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token no valido");
+        }
         return usuarioService.allUsersByNameAndLastName(nombre,apellidos);
     }
     @GetMapping(value="/usuarios/nombre/{nombre}")
-    public ResponseEntity listarPorNombre(@PathVariable  String nombre ){
+    public ResponseEntity listarPorNombre(@PathVariable  String nombre , @RequestHeader(value="Authorization") String token){
+        if(jwtUtil.getKey(token) == null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token no valido");
+        }
         return usuarioService.allUsersByName(nombre);
     }
     @GetMapping(value="/usuarios/apellidos/{apellidos}")
-    public ResponseEntity listarPorapellidos(@PathVariable  String apellidos ){
+    public ResponseEntity listarPorapellidos(@PathVariable  String apellidos , @RequestHeader(value="Authorization") String token){
+        if(jwtUtil.getKey(token) == null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token no valido");
+        }
         return usuarioService.allUsersByLastName(apellidos);
     }
     @PutMapping(value = "/usuario/{id}")
-    public ResponseEntity editarUsuario(@PathVariable Long id,@RequestBody @Valid Usuario usuario){
+    public ResponseEntity editarUsuario(@PathVariable Long id,@RequestBody @Valid Usuario usuario, @RequestHeader(value="Authorization") String token){
+        if(jwtUtil.getKey(token) == null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token no valido");
+        }
         return usuarioService.editUser(id,usuario);
     }
     @DeleteMapping (value = "/usuario/{id}")
-    public ResponseEntity eliminarUsuario(@PathVariable Long id) {
+    public ResponseEntity eliminarUsuario(@PathVariable Long id, @RequestHeader(value="Authorization") String token) {
+        if(jwtUtil.getKey(token) == null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token no valido");
+        }
         return usuarioService.deleteUser(id);
     }
     @ResponseStatus(HttpStatus.BAD_REQUEST)
